+++ src/maps/geo-utils/operators.ts
@@ -1,0 +1,52 @@
+import * as turf from "@turf/turf";
+import type {
+    Feature,
+    FeatureCollection,
+    MultiPolygon,
+    Polygon,
+} from "geojson";
+
+import { BLANK_GEOJSON } from "@/maps/api";
+
+export const safeUnion = (input: FeatureCollection<Polygon | MultiPolygon>) => {
+    if (input.features.length === 1) return input.features[0];
+    const union = turf.union(input);
+    if (union) return union;
+    throw new Error("No features");
+};
+
+export const holedMask = (
+    input:
+        | Feature<Polygon | MultiPolygon>
+        | FeatureCollection<Polygon | MultiPolygon>,
+) => {
+    return turf.difference(
+        turf.featureCollection([
+            BLANK_GEOJSON.features[0] as Feature<Polygon>,
+            "features" in input ? safeUnion(input) : input,
+        ]),
+    );
+};
+
+export const modifyMapData = (
+    mapData: FeatureCollection<Polygon | MultiPolygon>,
+    modifications:
+        | FeatureCollection<Polygon | MultiPolygon>
+        | Feature<Polygon | MultiPolygon>,
+    withinModifications: boolean,
+) => {
+    const safeModifications =
+        "features" in modifications ? safeUnion(modifications) : modifications;
+
+    if (withinModifications) {
+        return turf.intersect(
+            turf.featureCollection([safeUnion(mapData), safeModifications]),
+        );
+    }
+    return turf.intersect(
+        turf.featureCollection([
+            safeUnion(mapData),
+            holedMask(safeModifications)!,
+        ]),
+    );
+};
